back_end/requirements.txt
Django==2.2.6
django-cors-headers==3.1.1
django-environ==0.4.5
django-filter==2.2.0
djangorestframework==3.10.3
